Moved Server Logic To Own File
